// lib/constants/geo/world-countries.generated.ts
export const WORLD_COUNTRIES: WorldCountryRow[] = [
  {
    "name": "Afghanistan",
    "isoCode": "AF",
    "phoneCode": "+93"
  },
// ...
] as const;

export const WORLD_COUNTRY_NAMES: string[] = [
  ...WORLD_COUNTRIES.map((c) => c.name),
  "Other",
];

export const WORLD_COUNTRY_BY_NAME = new Map(WORLD_COUNTRIES.map((c) => [c.name, c]));